requirements.txt
streamlit
scikit-learn
joblib
pandas
requests
python-dotenv
